代码重构

+++ Camera/PerspectiveCamera.h
@@ -11,17 +11,17 @@ private:
     Vec3 vertical;
 
 public:
-    PerspectiveCamera(Point3 look_from, Point3 look_at, Vec3 up, double fov, double aspect_ratio);
+    PerspectiveCamera(Point3 look_from, Point3 look_at, Vec3 up, float fov, float aspect_ratio);
 
-    virtual Ray GetRay(const double &s, const double &t) const override;
+    virtual Ray GetRay(const float &s, const float &t) const override;
 };
 
-PerspectiveCamera::PerspectiveCamera(Point3 look_from, Point3 look_at, Vec3 up, double fov, double aspect_ratio) {
-    double theta = Degrees2Radians(fov);
-    double h = std::tan(theta / 2);
+PerspectiveCamera::PerspectiveCamera(Point3 look_from, Point3 look_at, Vec3 up, float fov, float aspect_ratio) {
+    float theta = Degrees2Radians(fov);
+    float h = std::tan(theta / 2);
 
-    double viewport_height = 2.0 * h;
-    double viewport_width = aspect_ratio * viewport_height;
+    float viewport_height = 2.0 * h;
+    float viewport_width = aspect_ratio * viewport_height;
 
     Vec3 w = (look_from - look_at).Normalize();
     Vec3 u = up.Cross(w).Normalize();
@@ -33,6 +33,6 @@ PerspectiveCamera::PerspectiveCamera(Point3 look_from, Point3 look_at, Vec3 up, 
     start = origin - horizontal / 2 - vertical / 2 - w;
 }
 
-Ray PerspectiveCamera::GetRay(const double &s, const double &t) const {
+Ray PerspectiveCamera::GetRay(const float &s, const float &t) const {
     return Ray(origin, start + s * horizontal + t * vertical - origin);
 }
