+++ Sampler/TrapezoidalSampler.h
@@ -13,14 +13,14 @@ private:
 Vec2 TrapezoidalSampler::CastRayByDistribution(int x, int y) {
     index = index>=4?0:index;
 
-    double r1 = 2 * RandomDouble();
-    double r2 = 2 * RandomDouble();
+    float r1 = 2 * RandomFloat();
+    float r2 = 2 * RandomFloat();
 
-    double dx = r1 < 1 ? sqrt(r1) - 1 : 1 - sqrt(2 - r1);
-    double dy = r2 < 1 ? sqrt(r2) - 1 : 1 - sqrt(2 - r2);
+    float dx = r1 < 1 ? sqrt(r1) - 1 : 1 - sqrt(2 - r1);
+    float dy = r2 < 1 ? sqrt(r2) - 1 : 1 - sqrt(2 - r2);
 
-    double s = static_cast<double>(x) + (dx + index % 2 + 0.5) / 2;
-    double t = static_cast<double>(y) + (dy + index / 2 + 0.5) / 2;
+    float s = static_cast<float>(x) + (dx + index % 2 + 0.5) / 2;
+    float t = static_cast<float>(y) + (dy + index / 2 + 0.5) / 2;
 
     index++;
