+++ Tool/Vector2.h
@@ -3,30 +3,30 @@
 struct Vector2 {
     union {
         struct {
-            double x, y;
+            float x, y;
         };
         struct {
-            double u, v;
+            float u, v;
         };
         struct {
-            double s, t;
+            float s, t;
         };
-        double data[2];
+        float data[2];
     };
 
     Vector2() : data{0, 0} {}
 
-    Vector2(double _value) : data{_value, _value} {}
+    Vector2(float _value) : data{_value, _value} {}
 
-    Vector2(double _x, double _y) : data{_x, _y} {}
+    Vector2(float _x, float _y) : data{_x, _y} {}
 
     //矢量取反
     Vector2 operator-() const { return {-x, -y}; }
 
     //矢量索引
-    double operator[](int i) const { return data[i]; }
+    float operator[](int i) const { return data[i]; }
 
-    double &operator[](int i) { return data[i]; }
+    float &operator[](int i) { return data[i]; }
 
     //矢量四则运算
     Vector2 operator+(const Vector2 &v) const {
