+++ Tool/Vector4.h
@@ -5,13 +5,13 @@
 struct Vector4{
     union{
         struct {
-            double x,y,z,w;
+            float x,y,z,w;
         };
-        double data[4];
+        float data[4];
     };
 
     Vector4():data{0,0,0,0}{}
-    Vector4(double _x,double _y,double _z,double _w):data{_x,_y,_z,_w}{}
+    Vector4(float _x,float _y,float _z,float _w):data{_x,_y,_z,_w}{}
     Vector4(Vector3 _v):data{_v.x,_v.y,_v.z,1}{}
 
 };
